+++ lib/src/request/request.dart
@@ -5,6 +5,7 @@ part 'auth/refresh_token_request_dto.dart';
 part 'auth/register_request_dto.dart';
 part 'payment/payment_request_dto.dart';
 part 'payment/payment_update_request_dto.dart';
+part 'project/project_label_request_dto.dart';
 part 'project/project_request_dto.dart';
 part 'request.freezed.dart';
 part 'request.g.dart';