+++ lib/src/response/response.dart
@@ -4,6 +4,7 @@ part 'auth/login_response_dto.dart';
 part 'error/failure_model.dart';
 part 'payment/payment_response_dto.dart';
 part 'project/project_detail_response.dart';
+part 'project/project_label_response.dart';
 part 'project/project_response.dart';
 part 'response.freezed.dart';
 part 'response.g.dart';